Fixing Typescript errors for v4
